debugged for voxel pipeline

+++ scripts/functions/pytorchtools.py
@@ -115,14 +115,14 @@ def collate_voxels(batch):
     for i in range(len(xx)):
         x = xx[i]
         y = yy[i]
-        target = torch.zeros(x_max, y_max, z_max, 4).type(torch.float16)
-        target[:x.shape[0],:x.shape[1],:x.shape[2],:] = x.type(torch.float16)
+        target = torch.zeros(x_max, y_max, z_max, 4)
+        target[:x.shape[0],:x.shape[1],:x.shape[2],:] = x
         xx_pad.append(target.to(device))
-        yy_pad.append(y.float().to(device))
+        yy_pad.append(y.to(device))
         #target = target + (0.1**0.5)*torch.randn(target.shape)
 
 
-    yy_pad = torch.stack(yy).type(torch.float16).to(device)
-    xx_pad = torch.stack(xx_pad).to(device)
+    yy_pad = torch.stack(yy).type(torch.float).to(device)
+    xx_pad = torch.stack(xx_pad).type(torch.float).to(device)
 
     return xx_pad, yy_pad, x_lens, y_lens
